feat: further styling updates, adding an auto-scroll-to-top
