+++ src/styles/footer.css
@@ -25,12 +25,12 @@ footer {
 
 .mx-icon {
   width: clamp(50px, 100px, 120px);
-  opacity: 0.7;
+  opacity: 1;
   transition: ease 0.4s;
 }
 
 .mx-icon:hover {
-  opacity: 1;
+  opacity: 0.7;
   transform: translate(0, -8px);
   transition: ease 0.4s;
 }
