+++ src/styles/header.css
@@ -7,13 +7,16 @@ header {
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  margin: 0;
   width: 100vw;
-  background-color: rgba(255,255,255,0.7);
+  background-color: rgba(255, 255, 255, 0.7);
   .logo {
     padding: 20px;
-    max-width: 8vw;
-    min-width: 75px;
+    width: clamp(75px, 8vw, 150px);
     transition: ease 0.4s;
+    align-items: center;
+    justify-content: center;
+    margin: 0;
   }
   .logo:hover {
     opacity: 0.7;
