+++ src/styles/home.css
@@ -6,7 +6,7 @@
   margin: 0;
   animation: fadeInAnimation ease 1s;
   animation-iteration-count: 1;
-  animation-fill-mode: forwards;
+  animation-fill-mode: forwards
 }
 @keyframes fadeInAnimation {
   0% {
@@ -24,10 +24,8 @@ body {
 }
 
 main {
-  /* This keeps sufficient space for the header at the top: */
-  padding-top: clamp(75px, 12vh, 150px);
   /* This leaves sufficient space for the footer at the bottom: */
-  padding-bottom: 35vh;
+  padding-bottom: 10vh;
   line-height: 1.4;
 }
 
@@ -52,8 +50,7 @@ main {
 /* Styling specific to the homepage */
 
 .power-div {
-  /* Additional white space for homepage only */
-  padding-top: clamp(50px, 5vh, 100px);
+  padding-top: 15vh;
   /* Specifying width etc on a div-level to allow for variations */
   width: 75vw;
   display: block;
@@ -113,7 +110,8 @@ main {
     display: flex;
     flex-direction: inline-flex;
     justify-content: space-between;
-    margin: 20px 2vw;
+    margin: 0 2vw;
+    padding: 30px 0;
     border-bottom: 1px solid #cccccc;
   }
 
@@ -136,7 +134,12 @@ main {
     width: clamp(50%, 65%, 75%);
     display: flex;
     flex-direction: column;
-    margin-bottom: 20px;
+  }
+
+  .job-button-div {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
   }
 
   .job-button {
@@ -157,6 +160,11 @@ main {
     background-color: #4f5052;
     cursor: pointer;
   }
+}
+
+.mountain-img {
+  width: 100vw;
+  margin-top: 5vh;
 }
 
 /* Adjustments for smaller computer screens: */
@@ -193,7 +201,7 @@ main {
 /* Adjustments for tablets: */
 @media only screen and (max-width: 768px) {
   main {
-    padding-bottom: 25vh;
+    padding-bottom: 15vh;
   }
   .power-div {
     width: 85vw;
@@ -239,8 +247,8 @@ main {
       display: block;
     }
 
-    .job-button {
-      margin-bottom: 20px;
+    .job-item-details {
+      margin-bottom: 15px;
     }
   }
 }
